+++ tests/Rocksmith2014.DD.Tests/ComparerTests.fs
@@ -36,6 +36,18 @@ let maxSimilarityFastTests =
 [<Tests>]
 let comparerTests =
     testList "Same Element Count Tests" [
+        testCase "Count is zero for different notes" <| fun _ ->
+            let notes1 = [ Note(Time = 25, String = 0y, Fret = 1y)
+                           Note(Time = 50, String = 0y, Fret = 5y)
+                           Note(Time = 75, String = 1y, Fret = 9y) ]
+            let notes2 = [ Note(Time = 125, String = 0y, Fret = 2y)
+                           Note(Time = 175, String = 0y, Fret = 4y)
+                           Note(Time = 200, String = 1y, Fret = 8y) ]
+
+            let sameCount = getSameElementCount sameNote notes1 notes2
+
+            Expect.equal sameCount 0 "Count is correct"
+
         testCase "Correct for same notes" <| fun _ ->
             let notes1 = [ Note(Time = 25, String = 0y, Fret = 14y)
                            Note(Time = 50, String = 0y, Fret = 13y)
@@ -299,4 +311,45 @@ let comparerTests =
             getSameElementCount sameNote notes1 notes2 |> ignore
 
             Expect.isTrue true "An eternity has not passed by"
+
+        // [ 4   1 8 8 8 ]
+        // [ 5 0 1 8 8 8 ]
+        testCase "Correct when starting note is different and different note in between" <| fun _ ->
+            let notes1 = [ Note(Time = 0, String = 0y, Fret = 4y)
+                           Note(Time = 1, String = 0y, Fret = 1y)
+                           Note(Time = 2, String = 0y, Fret = 8y)
+                           Note(Time = 50, String = 0y, Fret = 8y)
+                           Note(Time = 75, String = 0y, Fret = 8y) ]
+            let notes2 = [ Note(Time = 100, String = 0y, Fret = 5y)
+                           Note(Time = 100, String = 0y, Fret = 0y)
+                           Note(Time = 125, String = 0y, Fret = 1y)
+                           Note(Time = 175, String = 0y, Fret = 8y)
+                           Note(Time = 200, String = 0y, Fret = 8y)
+                           Note(Time = 225, String = 0y, Fret = 8y) ]
+
+            let sameCount = getSameElementCount sameNote notes1 notes2
+
+            Expect.equal sameCount 4 "Count is correct"
+
+        // [ 5             1 2 3 ]
+        // [ 5 0 0 0 0 0 0 1 2 3 ]
+        testCase "Correct when starting 6 different notes in between" <| fun _ ->
+            let notes1 = [ Note(Time = 0, String = 0y, Fret = 5y)
+                           Note(Time = 1, String = 0y, Fret = 1y)
+                           Note(Time = 2, String = 0y, Fret = 2y)
+                           Note(Time = 50, String = 0y, Fret = 3y) ]
+            let notes2 = [ Note(Time = 50, String = 0y, Fret = 5y)
+                           Note(Time = 100, String = 0y, Fret = 0y)
+                           Note(Time = 110, String = 0y, Fret = 0y)
+                           Note(Time = 120, String = 0y, Fret = 0y)
+                           Note(Time = 130, String = 0y, Fret = 0y)
+                           Note(Time = 140, String = 0y, Fret = 0y)
+                           Note(Time = 150, String = 0y, Fret = 0y)
+                           Note(Time = 200, String = 0y, Fret = 1y)
+                           Note(Time = 210, String = 0y, Fret = 2y)
+                           Note(Time = 220, String = 0y, Fret = 3y) ]
+
+            let sameCount = getSameElementCount sameNote notes1 notes2
+
+            Expect.equal sameCount 4 "Count is correct"
     ]