+++ tests/DLCBuilder.Domain.Tests/MiscTests.fs
@@ -17,6 +17,7 @@ let testProject =
       Year = DateTime.Now.Year
       AlbumArtFile = ""
       AudioFile = AudioFile.Empty
+      AudioFileLength = None
       AudioPreviewFile = AudioFile.Empty
       AudioPreviewStartTime = None
       PitchShift = None