+++ samples/DLCBuilder/RELEASE_NOTES.md
@@ -1,7 +1,8 @@
-## 3.1.0 (Released 2024-xx-xx)
+## 3.1.0 (Released 2024-06-01)
 
 - Added validation that all characters used in the lyrics are found in a custom font.
 - The tone keys lists for arrangements are now updated when the project is built.
+- The length of the audio file is now saved to the project file.
 - Fixed an issue where the DD generation would in rare cases get stuck when calulating phrase similarities.
 - Fixed an issue where PSARC import would fail in rare cases.